build: leave egui 0.29 behind for 0.36
Two of the four advisories deny.toml ignored were live vulnerabilities
whose only fix was a breaking quick-xml release, which meant bumping
accesskit, which meant bumping egui. So the stack the repo had been
holding at 0.29 for two releases moves to 0.36, and the ignore list goes
from four entries to one: RUSTSEC-2024-0436 (paste) left the graph and
RUSTSEC-2026-0194/0195 (quick-xml via zbus_xml/atspi) went with the
accesskit bump. Only the ttf-parser unmaintained notice remains.

The dependency wall came first: wgpu-hal 30 wants gpu-allocator built
against windows 0.62, but gpu-allocator accepts a range spanning 0.53 to
0.62 and the old lockfile pinned it to the 0.56 that trash requires, so
it compiled its D3D12 calls against one windows crate and handed them to
another. Regenerating the lockfile resolves the flexible edge upward and
the two versions coexist, as they are meant to.

Three subsystems were redesigned rather than renamed:

- eframe::App splits update into logic(ctx, frame) and ui(ui, frame).
  Polling the background workers belongs in logic, which also runs while
  the window is hidden; drawing takes the Ui.
- SidePanel and TopBottomPanel collapse into Panel::top/bottom/left/right
  taking that Ui, with default_size/min_size in place of the per-axis
  builders.
- The menu system is rebuilt. menu::bar becomes MenuBar, and the styling
  that used to be applied inside the bar (because set_menu_style
  overwrote anything set on the way in) is now a StyleModifier handed to
  MenuBar::style, which egui applies inside the scope for us. It calls
  egui menu_style first, because a modifier replaces the default rather
  than adding to it. close_menu becomes ui.close(), and the bar asks for
  CloseOnClickOutside: 0.32 made CloseOnClick the default, which would
  dismiss the menu under the size choices and the three view toggles that
  are deliberately flippable several at a time.

Two behaviour changes under the API mattered more than the renames:

- A Frame counts its stroke width as padding now. A 1px border on a pane
  frame insets its content to 13px, which broke the one column the
  spacing scale exists to keep -- every_pane_rules_off_its_heading_at_the_same_inset
  caught it at exactly 13.0. Panel frames therefore carry no stroke; the
  line between panes is the Panel separator, painted outside the content
  and already coloured from widgets.noninteractive.bg_stroke, which is
  the same palette border the frames used to stroke. The menu bar needed
  it for the second reason too: with a stroke the highlight could no
  longer fill the strip.
- Margin, CornerRadius and Shadow are whole pixels. The scale stays f32
  for add_space and Vec2 and crosses into Margin through theme::px, one
  conversion rather than an as-cast at sixteen call sites.

Also mechanical: Rounding to CornerRadius, Frame::none to Frame::NONE, an
explicit StrokeKind::Middle on every painted rect stroke (Middle is what
0.29 painted), ColorImage gaining source_size, ViewportCommand::Screenshot
taking UserData, Context::style to global_style, screen_rect to
viewport_rect, wants_keyboard_input to egui_wants_keyboard_input, and
Context::run to run_ui in the test harness.

Three test-visible behaviour changes, each fixed rather than papered over:

- epaint now panics if a frame is dropped with texture deltas unapplied.
  The tests have no renderer, so they discard them explicitly.
- egui_extras puts a column resize handle on the next column's leading
  edge rather than one item-spacing past the previous cell, so the drag
  helper derives its target from the pair of cells instead of a spacing
  constant, and aims one pixel inside the line so the next header's own
  click sense cannot swallow the press.
- Text layout can come out fractionally wider than the sum of its glyph
  advances, which is what truncate_for_width was assuming. The glyph scan
  is now the estimate and a single layout is the check, giving characters
  back until the label really fits.

The modal backdrop blur is softened while the code is open: the blur
width doubles to 440, which halves the downscale factor and so halves the
effective radius. At 220 a 1280px window blurred with a radius around 20
screen pixels, which stopped reading as the app behind the card.

CLAUDE.md and CONTRIBUTING.md are rewritten where they described the old
API, and dependabot no longer ignores the egui stack above patch.

+++ src/gui/ui/probes.rs
@@ -42,7 +42,7 @@ pub(super) static TEST_MENU_BAR_RECTS: std::sync::Mutex<Vec<(String, egui::Rect)
 /// menu bar has to override it — a rounded pill under a menu name reads
 /// as a floating button rather than as part of the bar.
 #[cfg(test)]
-pub(super) static TEST_MENU_BAR_ROUNDING: std::sync::Mutex<Vec<(String, f32)>> =
+pub(super) static TEST_MENU_BAR_ROUNDING: std::sync::Mutex<Vec<(String, u8)>> =
     std::sync::Mutex::new(Vec::new());
 
 /// `(label, x of the first glyph)` for every table column header, of
